Fixed the missing audio tone during reset.
This was corrected by initializing the pia_1 ports to emulate the effect of the pia's internal pullups at power up.  Additionally, the cpu is halted for 1.1ms after reset to match the duration of the reset pulse after power up.  This combination correctly produces the tone that occurs when the game is powered on.


git-svn-id: svn://svn.code.sf.net/p/pinmame/code/trunk@2987 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf