Fix AlvinG solenoids not firing consistently (most notably breaking Mystery Castle)
Cause: In SVN rev 4253, a hack was removed by slowing the interrupts-per-VBLANK interval from 60 to 10 (so, 3600hz to 600hz)   The solenoid smoothing intervals were then INCREEASED by a factor of 10, but the VBLANK interval actually SLOWED by a factor of 6, so very excessive "smoothing" was masking solenoid activity.  

I suspect the intent was to maintain the existing smoothing rate, but the math got a bit inverted. 

The old working value was so low (4 at 3600hz = 900hz) as to essentially not be "smoothing" anything at all, so setting it to 1 here to restore functionality. 


git-svn-id: svn://svn.code.sf.net/p/pinmame/code/trunk@4934 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf